requirements.txt
Flask==2.3.2
nltk==3.8.1
scikit-learn==1.3.0
PyPDF2==3.0.1
google-generativeai==0.5.1
